backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
python-multipart=